assistant/orchestrator_requirements.txt
# assistant/orchestrator_requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
httpx==0.27.0
loguru==0.7.2
langchain==0.2.14
langchain-community==0.2.5
langgraph==0.1.1
pika==1.3.2
pydantic-settings==2.3.4
webrtcvad-wheels==2.0.10.post2
pvporcupine==1.9.5
pyaudio==0.2.14
soundfile==0.12.1
neo4j==5.22.0
spacy==3.7.5
langchain-chroma==0.1.1
unstructured==0.14.9
